requirements.txt
ctransformers == 0.2.5
sentence-transformers == 3.2.0
pinecone-client == 5.0.1
langchain == 0.2.16
langchain_pinecone == 0.1.3
langchain_openai == 0.1.25
langchain-community == 0.2.17
langchain-huggingface == 0.0.3
pinecone == 5.3.1
flask == 3.0.3
pypdf == 5.0.1
python-dotenv == 1.0.1
-e .